Fix bugs; Use checkout and reference number differently

+++ backend/src/main/java/com/gadget/rental/rental/RentalGadgetDTO.java
@@ -1,5 +1,6 @@
 package com.gadget.rental.rental;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 import jakarta.validation.constraints.Max;
@@ -12,6 +13,6 @@ public record RentalGadgetDTO(
         @Size(min = 1, max = 3) MultipartFile[] images,
         @NotEmpty String name,
         @NotEmpty ZonedDateTime createdAt,
-        @NotEmpty double price,
+        @NotEmpty BigDecimal price,
         @NotEmpty @Max(value = 2000) String description) {
 }
